simple: Remove check for hop-by-hop headers
This check assumed that the headers would be lowercase, which is not
(yet) guaranteed by the WAS protocol or the implementation itself.
Fortunately this check is redundant since the role of a WAS server
requires that hop-by-hop headers will not be sent to it (assuming a
well-behaved WAS client), as they are not to be forwarded by
proxies (RFC2616, 13.5.1).
http_header_is_hop_by_hop also checked for Content-Length, which
is explicitly excluded by the WAS protocol (and replaced by a
LENGTH message).